Progress 5/19/2021
Making progress. The cacheIndexes are now getting updated in sync with
the accessInputProcessBlock transitions under goToEvent. Finally.

There are seg faults in a couple unit tests now that I need to debug.

And I still need to look at skipEvents, nextFile and previousFile.
I think there is still an issue there. I don't think the file
transitions are being triggered in every case when the file changes
which means the events cannot use the ProcessBlockCache. Still need
to get that going. Some of this is pre-existing issues with the
way the code that navigates around in the files worked...
Now it is important that the file transitions are firing when
necessary.

It was working OK for a simple forward iteration through the input.