Progress 5/21/2021
Still need to remove some print statements!!!!
I want to take one more look Monday, but maybe
I am done with IOPool/Input. There are now
working tests for skipEvents and goToEvent
and that was the last problematic part in
IOPool/Input. Hopefully IOPool/Output and
FWCore/Common are easier. There is already
tests for a lot of it already.

At least I still need a drop on output tests
and maybe a SubProcess test... I am not sure
how much the existing tests cover that.