Progress 5/27/2021
Unit test extensions. Lots of stepping through things
with the debugger. Another day with no bugs found!

Tomorrow need to add tests reading the 3 file test,
the drop on output test and the non-strict test. For
each test the cacheIndices. That should be sufficient for
OutputProcessBlockHelper.

Then SubProcessBlockHelper.

Then runTheMatrix, checkdeps and submit the PR.

All tomorrow? Monday or Tuesday is more likely.