Progress 5/28/2021
More unit tests, mostly related to filling the caches.

What is left.

1. Rerun the non-strict tests, temporarily edit ProductRegistry.cc
so those will work.

2. Test filling the caches in the non-strict tests. Add explicit checks
of the cache contents similar to what I was doing Friday. Maybe add
a new _cfg.py to read the output of the last step.

3. Add an EDAlias test. I think EDAlias should just work without any
special implementation, but it worth trying it.

4. Test SubProcessBlockHelper.cc. Hopefully this is not hard. Much
of it leverages other code elsewhere that is already tested and there
are already SubProcess tests in there where I tested SubProcess.cc.

5. runTheMatrix, checkdeps -a, code-checks again, code-format again

6. Skim thru the test code. But do not carefully review it like the
rest of it. You don't need tests of tests.

7. Look at my notes. Did I forget anything?

8. Submit the PR.