Progress 6/1/2021
This finishes OutputProcessBlockHelper. No bugs found again today.
A good sign. Lots of days with none. Last class to check is
SubProcessBlockHelper. Goal is to finish that one tomorrow!

Then an EDAlias test. It ought to just work. There is no new code
related to it.

Then runTheMatrix, checkdeps -a, code-checks, code-format.

Anything else? Maybe not.

Submit the PR.